Separate WorldMixin into WorldMixin and WorldMixinClient
Fix UltraMine compatibility.
